Utility to authorize EcoBee app
Creates EcoBee credential file (tokens.txt) for the first time, or if tokens expire.

This uses the PIN authorization method.  Once you have created an application in EcoBee portal (developer account), you must authorize the application in the "My Apps" section of the portal.  This will provide the pin requested by the utility.